fix: clipping error in bbepDrawSprite
In `bbepDrawSprite`, there are several calculations done to
avoid drawing parts of the sprite when outside of the display
area.

These calculations fail when using screen rotation, as those
checks are using the native width and height instead of the
calculated width and height updated when calling bbepSetRotation.

This commit also fixes a few additional issues when clipping the
maximum values.